+++ forui/lib/src/widgets/header/header.dart
@@ -36,7 +36,7 @@ sealed class FHeader extends StatelessWidget {
   const factory FHeader.nested({
     Widget title,
     FHeaderStyle? style,
-    TextAlign titleAlignment,
+    AlignmentGeometry titleAlignment,
     List<Widget> prefixes,
     List<Widget> suffixes,
     Key? key,